// enums/editor.enums.ts
export enum ActionKey {
  ADD_CARD = "add_card",
  REMOVE_CARD = "remove_card",
  EDIT_LABELS = "edit_labels",
  RESET_SPREAD = "reset_spread",
  SAVE_SPREAD = "save_spread",
  LOAD_SPREAD = "load_spread",
  EXIT = "exit",
  TOGGLE_GUIDELINES = "toggle_guidelines",
  TOGGLE_ROTATION = "toggle_rotation",
  TOGGLE_LABELS = "toggle_labels",
  TOGGLE_SEQUENCE = "toggle_sequence",
}